refactor: update imports to use UserType from entities in test files
