+++ src/test/kotlin/data/local/UserDataSourceTest.kt
@@ -7,7 +7,7 @@ import kotlinx.coroutines.test.runTest
 import org.baghdad.data.datasource.DataSource
 import org.baghdad.data.dto.UserDto
 import org.baghdad.data.local.UserDataSource
-import org.baghdad.logic.model.enums.UserType
+import org.baghdad.logic.model.entities.UserType
 import org.baghdad.logic.model.exceptions.UserNotFoundException
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
@@ -24,7 +24,7 @@ class UserDataSourceTest {
         name = "name",
         username = "username",
         hashedPassword = "hashedPassword",
-        type = UserType.Admin.toString(),
+        type = UserType.Admin.name,
     )
 
     @BeforeEach
