+++ src/test/kotlin/logic/usecase/project/EditProjectUseCaseTest.kt
@@ -4,7 +4,7 @@ import helpers.authentication.createUserHelper
 import io.mockk.*
 import kotlinx.coroutines.test.runTest
 import org.baghdad.logic.model.entities.ProjectEntity
-import org.baghdad.logic.model.enums.UserType
+import org.baghdad.logic.model.entities.UserType
 import org.baghdad.logic.model.exceptions.AccessDeniedException
 import org.baghdad.logic.model.exceptions.EmptyProjectNameException
 import org.baghdad.logic.repositories.AuditRepository
