+++ src/test/kotlin/logic/usecase/task/CreateTaskUseCaseTest.kt
@@ -2,15 +2,11 @@ package logic.usecase.task
 
 import com.google.common.truth.Truth.assertThat
 import helpers.task.TaskEntityTestData
-import io.mockk.Called
-import io.mockk.coVerify
-import io.mockk.mockk
-import io.mockk.slot
-import io.mockk.verify
+import io.mockk.*
 import kotlinx.coroutines.test.runTest
 import org.baghdad.logic.model.entities.AuditLogEntity
 import org.baghdad.logic.model.entities.UserEntity
-import org.baghdad.logic.model.enums.UserType
+import org.baghdad.logic.model.entities.UserType
 import org.baghdad.logic.model.exceptions.TaskWithMissingDescriptionException
 import org.baghdad.logic.model.exceptions.TaskWithMissingTitleException
 import org.baghdad.logic.repositories.AuditRepository
@@ -19,8 +15,7 @@ import org.baghdad.logic.repositories.UserRepository
 import org.baghdad.logic.usecase.task.CreateTaskUseCase
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.assertThrows
-import java.util.UUID
-import kotlin.jvm.java
+import java.util.*
 import kotlin.test.Test
 
 class CreateTaskUseCaseTest {
