+++ src/test/kotlin/logic/usecase/user/UserValidatorUseCaseTest.kt
@@ -4,7 +4,7 @@ import helpers.authentication.createUserHelper
 import io.mockk.coEvery
 import io.mockk.mockk
 import kotlinx.coroutines.test.runTest
-import org.baghdad.logic.model.enums.UserType
+import org.baghdad.logic.model.entities.UserType
 import org.baghdad.logic.model.exceptions.*
 import org.baghdad.logic.repositories.UserRepository
 import org.baghdad.logic.usecase.user.UserValidatorUseCase
