+++ src/test/kotlin/presentation/user/CreateUserUITest.kt
@@ -7,17 +7,13 @@ import io.mockk.verify
 import kotlinx.coroutines.test.runTest
 import org.baghdad.logic.manager.SessionManager
 import org.baghdad.logic.model.entities.UserEntity
-import org.baghdad.logic.model.enums.UserType
-import org.baghdad.logic.model.exceptions.InvalidNameException
-import org.baghdad.logic.model.exceptions.InvalidPasswordException
-import org.baghdad.logic.model.exceptions.InvalidUsernameException
-import org.baghdad.logic.model.exceptions.UnauthorizedException
-import org.baghdad.logic.model.exceptions.UserAlreadyExistsException
+import org.baghdad.logic.model.entities.UserType
+import org.baghdad.logic.model.exceptions.*
 import org.baghdad.logic.usecase.user.CreateUserUseCase
 import org.baghdad.presentation.input.Reader
 import org.baghdad.presentation.output.Viewer
 import org.baghdad.presentation.user.CreateUserUI
-import java.util.UUID
+import java.util.*
 import kotlin.test.BeforeTest
 import kotlin.test.Test
